% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralneuro_adapet}
\alias{admiralneuro_adapet}
\title{Amyloid PET Scan Analysis Dataset - Neuro}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 83 rows and 50 columns.
}
\usage{
admiralneuro_adapet
}
\description{
An updated ADaM ADAPET dataset using NV, AG, SUPPNV, ADSL
}
\seealso{
Other datasets:
\code{\link{admiralneuro_adtpet}},
\code{\link{adsl_neuro}}
}
\concept{datasets}
\keyword{datasets}
