% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralneuro_adpet}
\alias{admiralneuro_adpet}
\title{PET Scan Analysis Dataset - Neuro}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 132 rows and 43 columns.
}
\usage{
admiralneuro_adpet
}
\description{
An updated ADaM ADPET dataset using NV and adsl
}
\seealso{
Other datasets: 
\code{\link{ag_neuro}},
\code{\link{dm_neuro}},
\code{\link{nv_neuro}},
\code{\link{suppnv_neuro}}
}
\concept{datasets}
\keyword{datasets}
