% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ag_neuro}
\alias{ag_neuro}
\title{Procedure Agents for Nervous System Dataset}
\format{
A data frame with 12 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{AGSEQ}{Sequence Number}
\item{AGTRT}{Tracer Name}
\item{AGCAT}{Tracer Category}
\item{AGDOSE}{Dose per Adminstration}
\item{AGDOSEU}{Dose Units}
\item{AGROUTE}{Route of Administration}
\item{AGLNKID}{Link to Subject's NV Domain Record}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{AGSTDTC}{Date/Time of Administration}
}
}
\source{
Constructed using \code{nv_neuro} from \code{{admiralneuro}} package
}
\usage{
ag_neuro
}
\description{
A SDTM AG domain dataset containing procedure agents for nervous system
}
\details{
Procedure Agents for Nervous System Dataset

A SDTM AG domain dataset
}
\seealso{
Other datasets: 
\code{\link{dm_neuro}},
\code{\link{nv_neuro}},
\code{\link{suppnv_neuro}}
}
\concept{datasets}
\keyword{keyword}
