% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suppnv_neuro}
\alias{suppnv_neuro}
\title{Supplemental Nervous System Findings Dataset}
\format{
A data frame with 8 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{RDOMAIN}{Related Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{IDVAR}{Identifying Variable}
\item{IDVARVAL}{Identifying Variable Value}
\item{QNAM}{Qualifier Variable Name}
\item{QLABEL}{Qualifier Variable Label}
\item{QVAL}{Data Value}
\item{QORIG}{Origin}
\item{QEVAL}{Evaluator}
}
}
\source{
Constructed using \code{nv_neuro} from \code{{admiralneuro}} package
}
\usage{
suppnv_neuro
}
\description{
A SDTM SUPPNV domain dataset containing reference regions used for SUVR calculation
}
\details{
Supplemental Nervous System Findings Dataset

A SDTM SUPPNV domain dataset
}
\seealso{
Other datasets: 
\code{\link{admiralneuro_adapet}},
\code{\link{admiralneuro_adtpet}},
\code{\link{adsl_neuro}},
\code{\link{ag_neuro}},
\code{\link{dm_neuro}},
\code{\link{nv_neuro}}
}
\concept{datasets}
\keyword{datasets}
