% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralneuro-package.R
\docType{package}
\name{admiralneuro-package}
\alias{admiralneuro}
\alias{admiralneuro-package}
\title{admiralneuro: Neuroscience Extension Package for ADaM in 'R' Asset Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Programming neuroscience specific Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in 'R'. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam}). This package extends the 'admiral' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiralneuro/}
  \item \url{https://github.com/pharmaverse/admiralneuro}
  \item Report bugs at \url{https://github.com/pharmaverse/admiralneuro/issues}
}

}
\author{
\strong{Maintainer}: Jian Wang \email{wang_jian_wj@lilly.com} (\href{https://orcid.org/0009-0002-4677-3781}{ORCID})

Authors:
\itemize{
  \item Jian Wang \email{wang_jian_wj@lilly.com} (\href{https://orcid.org/0009-0002-4677-3781}{ORCID})
  \item Miles Almond (\href{https://orcid.org/0009-0007-1784-0355}{ORCID})
  \item Xiao Chen (\href{https://orcid.org/0009-0000-6959-5151}{ORCID})
  \item Fanny Gautier (\href{https://orcid.org/0009-0004-3581-0131}{ORCID})
  \item Gayatri G.
  \item Meilin Jiang (\href{https://orcid.org/0000-0003-4515-4567}{ORCID})
  \item Leena Khatri (\href{https://orcid.org/0000-0002-2268-4023}{ORCID})
  \item Edoardo Mancini (\href{https://orcid.org/0009-0006-4899-8641}{ORCID})
  \item Eric Nantz
  \item Lina Patil
  \item Chris Pelentrides
  \item Katie Withycombe (\href{https://orcid.org/0009-0005-5376-1932}{ORCID})
}

Other contributors:
\itemize{
  \item Eli Lilly and Company [copyright holder, funder]
  \item Cytel Inc. [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
