% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralneuro-package.R
\docType{package}
\name{admiralneuro-package}
\alias{admiralneuro}
\alias{admiralneuro-package}
\title{admiralneuro: Neuroscience Extension Package for ADaM in 'R' Asset Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for programming Clinical Data Standards Interchange Consortium (CDISC) compliant Analysis Data Model (ADaM) datasets in R. ADaM datasets are a mandatory part of any New Drug or Biologics License Application submitted to the United States Food and Drug Administration (FDA). Analysis derivations are implemented in accordance with the "Analysis Data Model Implementation Guide" (CDISC Analysis Data Model Team, 2021, \url{https://www.cdisc.org/standards/foundational/adam}). The package is an extension package of the 'admiral' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiralneuro}
  \item \url{https://github.com/pharmaverse/admiralneuro}
}

}
\author{
\strong{Maintainer}: Jian Wang \email{wang_jian_wj@lilly.com}

Authors:
\itemize{
  \item Chris Pelentrides
  \item Edoardo Mancini
  \item Eric Nantz
  \item Fanny Gautier (\href{https://orcid.org/0009-0004-3581-0131}{ORCID})
  \item Gayatri G.
  \item Leena Khatri
  \item Meilin Jiang (\href{https://orcid.org/0000-0003-4515-4567}{ORCID})
  \item Miles Almond
}

Other contributors:
\itemize{
  \item Eli Lilly [copyright holder, funder]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\concept{internal}
\keyword{internal}
