% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admiralneuro_adtpet}
\alias{admiralneuro_adtpet}
\title{Tau PET Scan Analysis Dataset - Neuro}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34 rows and 47 columns.
}
\usage{
admiralneuro_adtpet
}
\description{
An updated ADaM ADTPET dataset using NV, AG, SUPPNV, ADSL
}
\seealso{
Other datasets: 
\code{\link{admiralneuro_adapet}},
\code{\link{adsl_neuro}}
}
\concept{datasets}
\keyword{datasets}
