% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adsl_neuro}
\alias{adsl_neuro}
\title{Subject Level Analysis Dataset-updated}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 36 columns.
}
\usage{
adsl_neuro
}
\description{
An updated ADaM ADSL dataset with Alzheimer's Disease patients
}
\seealso{
Other datasets: 
\code{\link{admiralneuro_adpet}},
\code{\link{ag_neuro}},
\code{\link{dm_neuro}},
\code{\link{nv_neuro}},
\code{\link{suppnv_neuro}}
}
\concept{datasets}
\keyword{datasets}
