% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ag_neuro}
\alias{ag_neuro}
\title{Procedure Agents for Nervous System Dataset}
\format{
A data frame with 12 columns:
\describe{
\item{STUDYID}{Study Identifier}
\item{DOMAIN}{Domain Abbreviation}
\item{USUBJID}{Unique Subject Identifier}
\item{AGSEQ}{Sequence Number}
\item{AGTRT}{Reported Agent Name}
\item{AGCAT}{Category for Category}
\item{AGDOSE}{Dose per Administration}
\item{AGDOSEU}{Dose Units}
\item{AGROUTE}{Route of Administration}
\item{AGLNKID}{Link ID}
\item{VISITNUM}{Visit Number}
\item{VISIT}{Visit Name}
\item{AGSTDTC}{Start Date/Time of Agent}
}
}
\source{
Constructed using \code{nv_neuro} from \code{{admiralneuro}} package
}
\usage{
ag_neuro
}
\description{
A SDTM AG domain dataset containing procedure agents for nervous system
}
\details{
Procedure Agents for Nervous System Dataset

A SDTM AG domain dataset
}
\seealso{
Other datasets: 
\code{\link{admiralneuro_adpet}},
\code{\link{adsl_neuro}},
\code{\link{dm_neuro}},
\code{\link{nv_neuro}},
\code{\link{suppnv_neuro}}
}
\concept{datasets}
\keyword{datasets}
