% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm_neuro}
\alias{dm_neuro}
\title{Demographic Dataset - Neuro}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 26 columns.
}
\usage{
dm_neuro
}
\description{
An updated SDTM DM dataset subset with age appropriate Alzheimer's Disease patients
}
\seealso{
Other datasets: 
\code{\link{admiralneuro_adpet}},
\code{\link{ag_neuro}},
\code{\link{nv_neuro}},
\code{\link{suppnv_neuro}}
}
\concept{datasets}
\keyword{datasets}
